requirements.txt
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyQt5Designer==5.14.1
PyYAML==6.0
